* Cleanup logic package * Cleanup date text converter * Extract file endings to application.properties
