+++ src/main/resources/application.properties
@@ -2,6 +2,9 @@ application.name=@project.name@
 application.description=@project.description@
 photo-rename.ui.theme=theme-standard.css
 photo-rename.instructions=Just drop all photos of your camera, phone and WhatsApp into one directory. Copy this jar into the same directory and run it by double clicking. Then simply click "Start/Run".
-photo-rename.dateTimePattern=yyyy-MM-dd HH.mm.ss
-photo-rename.dateTimeLocaleLanguage=de
-photo-rename.dateTimeLocaleCountry=DE
+photo-rename.file.image.endings=JPG,JPEG,PNG,WebP,GIF,ICO,BMP,TIFF,PSD,PCX,CRW,CR2,NEF,ORF,RAF,RWL,SRW,ARW,DNG,X3F
+photo-rename.file.image.raw.endings=RAW,RW2
+photo-rename.file.video.endings=MOV,MP4,M4V,3G2,3GP,3GP
+photo-rename.dateTime.pattern=yyyy-MM-dd HH.mm.ss
+photo-rename.dateTime.locale.language=de
+photo-rename.dateTime.locale.country=DE
